Re-architect how the decoder will operate.
For the time being, the generator is broken and will output `OpcodeMap.g.cs` despite not being used anymore.

Instead of having knowledge of every known opcode and deciding later which one to run (based on CPU level, supported prefixes such as VEX, etc.), it will only have knowledge of the opcodes it is allowed to decode. So, for an 8086 decoder, [0F] will decode as `POP_CS`, but for 80186 and newer, it will decode as a two byte map escape prefix (see `OneBytePrefix.cs` and `TwoBytePrefix.cs`).

This makes supporting reused opcode bytes easy. For example,, the 80386 and 80486 would register `[0F 10]` as UMOV_EbGb, but the Pentium III and newer would instead register `MOVUPS_VxWx`.

In addition, in order to support quirks such as [F1] decoding as `LOCK:` on the 8086, prefixes are no longer recognized by their byte values, but instead by looking up the byte in an array. The output of the array for [F1] would be `OneBytePrefix.Lock` on the 8086, and `null` on the 80186 and newer.

This design is very flexible as it moves the logic of juggling the various CPUs. In the future, it might even be possible to support the NEC V20/V30's 8080 emulation mode by having the `BRKEM` opcode full-on swap the main CPU decoder instance to one supporting the 8080's opcode map. The `RETEM` 8080 opcode would then switch back to the original x86 decoder.